requirements.txt
streamlit
pandas
folium
streamlit-folium
sentence-transformers
deep-translator
streamlit-mic-recorder
fpdf2
SpeechRecognition